claude-agent-sdk/requirements.txt
claude-agent-sdk>=0.1.44
python-dotenv>=1.0.0
